% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_activation.R
\docType{data}
\name{activation}
\alias{activation}
\alias{values_activation}
\alias{values_brulee_activation}
\title{Activation functions between network layers}
\format{
An object of class \code{character} of length 5.

An object of class \code{character} of length 18.
}
\usage{
activation(values = values_activation, label = "Activation Function")

values_activation

values_brulee_activation
}
\arguments{
\item{values}{A character string of possible values. See \code{values_activation}
in examples below.}

\item{label}{A label to use for plotting and printing.}
}
\description{
Activation functions between network layers
}
\details{
This parameter is used in \code{parsnip} models for neural networks such as
\code{parsnip:::mlp()}.
}
\examples{
values_activation
activation()
values_brulee_activation
}
\keyword{datasets}
