% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_attention.R
\name{attention-param}
\alias{attention-param}
\alias{attention_type}
\alias{values_attention_type}
\alias{dropout_attn}
\alias{num_attn_heads}
\alias{num_attn_blocks}
\alias{num_attn_feat}
\alias{target_token}
\alias{normalization}
\alias{values_normalization}
\title{Parameters for attention-based tabular models}
\usage{
attention_type(values = values_attention_type)

values_attention_type

dropout_attn(range = c(0, 0.5), trans = NULL)

num_attn_heads(range = c(1L, 8L), trans = NULL)

num_attn_blocks(range = c(1L, 6L), trans = NULL)

num_attn_feat(range = c(8L, 64L), trans = NULL)

target_token(values = c(TRUE, FALSE))

normalization(values = values_normalization)

values_normalization
}
\arguments{
\item{values}{A character string of possible values. See
\code{values_attention_type} in examples below.}

\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
These parameters are used with attention-based (transformer) tabular models,
such as \code{brulee_saint()} and \code{tabular_auto_int()} when fit with the \code{brulee}
engine.
}
\details{
\itemize{
\item \code{attention_type()}: The type of attention mechanism to use.
\item \code{dropout_attn()}: The proportion of attention weights to randomly set to
zero during model training.
\item \code{num_attn_heads()}: The number of parallel attention mechanisms (heads) in
the multi-head attention layer. Multiple attention heads allow the model to
attend to different aspects of the input features simultaneously.
\item \code{num_attn_blocks()}: The number of sequential attention blocks in the model
architecture. Each attention block consists of a multi-head attention layer
followed by feed-forward layers.
\item \code{num_attn_feat()}: The dimensionality of the feature space used in the
attention mechanism.
\item \code{target_token()}: A logical to specify whether the SAINT model should
add a supervised token to the embeddings.
}
}
\examples{
values_attention_type
attention_type()
num_attn_heads()
target_token()
}
