% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_attention_type.R
\name{attention_type}
\alias{attention_type}
\alias{values_attention_type}
\title{Attention type for tabular transformers}
\usage{
attention_type(values = values_attention_type)

values_attention_type
}
\arguments{
\item{values}{A character string of possible values. See
\code{values_attention_type} in examples below.}
}
\description{
The type of attention mechanism to use.
}
\details{
Used as a tuning parameter for \code{brulee_saint()} in the \code{brulee} package.
}
\examples{
values_attention_type
attention_type()
}
