% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_cal_method.R
\docType{data}
\name{class_cal_method}
\alias{class_cal_method}
\alias{class_reg_method}
\alias{values_cal_cls}
\alias{values_cal_reg}
\title{Methods for model calibration}
\format{
An object of class \code{character} of length 6.

An object of class \code{character} of length 5.
}
\usage{
class_cal_method(values = values_cal_cls)

class_reg_method(values = values_cal_reg)

values_cal_cls

values_cal_reg
}
\arguments{
\item{values}{A character string of possible values. See \code{values_cal_cls}
and \code{values_cal_reg} (in examples below).}
}
\description{
Methods for model calibration
}
\details{
This parameter is used in postprocessing methods in the tailor package.
}
\examples{
values_cal_cls
class_cal_method()

values_cal_reg
class_reg_method()
}
\keyword{datasets}
