% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_engine_catboost.R
\name{l2_leaf_reg}
\alias{l2_leaf_reg}
\alias{max_leaves}
\title{Possible engine parameters for catboost}
\usage{
l2_leaf_reg(range = c(0, 1), trans = transform_log10())

max_leaves(range = c(10, 64), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
These parameters are auxiliary to tree-based models that use the "catboost"
engine. They correspond to tuning parameters that would be specified using
\code{set_engine("catboost", ...)}.
}
\details{
"catboost" is an available engine in the parsnip extension package
\href{https://bonsai.tidymodels.org/}{bonsai}

\code{max_leaves()} is only used when the grow policy is set to \code{"Lossguide"}.

For more information, see the \href{https://catboost.ai/docs/en/references/training-parameters/}{catboost webpage}.
}
\examples{

l2_leaf_reg()
max_leaves()

}
