% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraints}
\alias{constraints}
\alias{has_constraint}
\alias{add_parameter_constraint}
\alias{remove_parameter_constraint}
\alias{update_parameter_constraint}
\title{Parameter constraints}
\usage{
has_constraint(object)

add_parameter_constraint(object, .constraint = NULL)

remove_parameter_constraint(object)

update_parameter_constraint(object, .constraint = NULL)
}
\value{
\code{\link[=has_constraint]{has_constraint()}} returns a logical, all others return parameter
sets.
}
\description{
These helper functions can help manage constraints among parameters.
}
\examples{
no_prm <- parameters(lambda = penalty(), mixture())

has_constraint(no_prm)

constr_prm <- add_parameter_constraint(ex_prm, penalty < 0.01)
constr_prm
has_constraint(constr_prm)

remove_parameter_constraint(constr_prm)
update_parameter_constraint(constr_prm, log10(penalty) < -2 & mixture > 1 / 2)
}
