% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_engine_dbscan.R
\name{radius}
\alias{radius}
\alias{min_points}
\title{Parameters for possible engine parameters for dbscan}
\usage{
radius(range = c(-2, 2), trans = transform_log10())

min_points(range = c(3L, 50L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
These parameters are used for density-based clustering methods that use the
dbscan engine. They correspond to tuning parameters that would be specified using
\code{set_engine("dbscan", ...)}.
}
\details{
To use these, check \code{?tidyclust::db_clust} to see how they are used.
\itemize{
\item \code{radius()} controls the radius used to determine core-points and cluster assignments
\item \code{min_points()} controls the minimum number of nearby points to be considered a core-point (including the point itself)
}
}
\examples{
radius()
min_points()
}
