% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{default_parameters}
\alias{default_parameters}
\alias{values_activation}
\alias{values_weight_func}
\alias{values_surv_dist}
\alias{values_weight_scheme}
\alias{values_token}
\title{Default parameter values}
\format{A character vector of parameter values.}
\usage{
values_activation

values_weight_func

values_surv_dist

values_weight_scheme

values_token
}
\description{
These are default parameter value objects for parameters that take a \code{values}
argument. When optimizing over qualitative parameter values, you can subset
the default arguments listed here and supply the result to \code{values} if you
want to optimize a different set of \code{values}.
}
\keyword{datasets}
