% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_deg_free.R
\name{deg_free}
\alias{deg_free}
\title{Degrees of freedom (integer)}
\usage{
deg_free(range = c(1, 5), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
The number of degrees of freedom used for model parameters.
}
\details{
One context in which this parameter is used is spline basis functions.
}
\examples{
deg_free()
}
