% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_dropout_attn.R
\name{dropout_attn}
\alias{dropout_attn}
\title{Dropout rate for attention weights}
\usage{
dropout_attn(range = c(0, 0.5), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
The proportion of attention weights to randomly set to zero during model
training.
}
\details{
Used as a tuning parameter for \code{tabular_auto_int()} in the \pkg{tabular} package
when fit with the \code{brulee} engine.
}
\examples{
dropout_attn()
dropout_attn(range = c(0, 0.3))

}
