% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_dropout_last.R
\name{dropout_last}
\alias{dropout_last}
\title{Dropout rate for the final layer}
\usage{
dropout_last(range = c(0, 0.5), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
The proportion of final-layer units to randomly set to zero during model
training.
}
\details{
Used as a tuning parameter for \code{brulee_saint()} in the \code{brulee} package.
}
\examples{
dropout_last()
dropout_last(range = c(0, 0.3))

}
