% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dials_parameter.R
\name{extract_dials_parameter}
\alias{extract_dials_parameter}
\alias{extract_dials_parameter.parameters}
\alias{extract_dials_parameter.recipe}
\alias{extract_dials_parameter.model_spec}
\alias{extract_dials_parameter.workflow}
\title{Return a dials parameter object associated with parameters}
\usage{
extract_dials_parameter(x, id, ...)

\method{extract_dials_parameter}{parameters}(x, id, ...)

\method{extract_dials_parameter}{recipe}(x, id, ...)

\method{extract_dials_parameter}{model_spec}(x, id, ...)

\method{extract_dials_parameter}{workflow}(x, id, ...)
}
\arguments{
\item{x}{The results of a call to \code{\link[=parameters]{parameters()}}, a recipe, model
specification, or workflow.}

\item{id}{A single string for the \code{id} of the parameter.}

\item{...}{Not currently used.}
}
\value{
A \code{dials} parameter object.
}
\description{
\code{extract_dials_parameter()} can extract a single \code{dials} parameter object from
different types of objects (e.g. parameter sets, recipes, etc.).
}
\examples{
glmn_param <- parameters(lambda = penalty(), mixture())
extract_dials_parameter(glmn_param, "lambda")

# fails:
# extract_dials_parameter(glmn_param, "penalty")
}
