% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{get_batch_sizes}
\alias{get_batch_sizes}
\title{Get batch sizes}
\usage{
get_batch_sizes(object, x, frac = c(1/10, 1/3), ...)
}
\arguments{
\item{object}{A \code{param} object or a list of \code{param} objects.}

\item{x}{The predictor data. In some cases (see below) this should only
include numeric data.}

\item{frac}{A double for the fraction of the data to be used for the upper
bound. For \code{get_n_frac_range()} and \code{get_batch_sizes()}, a vector of two
fractional values are required.}

\item{...}{Other arguments to pass to the underlying parameter
finalizer functions. For example, for \code{get_rbf_range()}, the dots are passed
along to \code{\link[kernlab:sigest]{kernlab::sigest()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
