% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_set.R
\name{grid_assign_regular}
\alias{grid_assign_regular}
\alias{grid_assign_random}
\title{Assign a grid type to a parameter}
\usage{
grid_assign_regular(x, ...)

grid_assign_random(x, ...)
}
\arguments{
\item{x}{A \code{param_set}.}

\item{...}{Name-value pairs. The name matches the name of the parameter
you are modifying. For \code{grid_assign_regular()}, the value is the number of
\code{levels} to generate for that parameter. For \code{grid_assign_random()}, the
value is the \code{size} of each parameter to generate.}
}
\description{
These functions take a \code{param_set} and add meta data about how the grid
should be generated for each parameter in the set.
\itemize{
\item \code{grid_assign_regular()} states that the variables specified in \code{...}
should be generated in a \code{"regular"} fashion.
\item \code{grid_assign_random()} states that the variables specified in \code{...} should
be generated in a \code{"random"} fashion.
}
}
\details{
Neither of these functions generate the grid themselves, that is the job
of \code{\link[=grid_generate]{grid_generate()}}.

When an \code{id} is set in \code{\link[=attach_param]{attach_param()}} (or other similar \code{attach_*()}
functions), you \emph{must} specify the name of that parameter using the
convention \code{<name>..<id>}. Generally, \code{id} should not be set manually,
and is useful only when there are multiple parameters of the same name
in the set. See the examples for more information.
}
\examples{

# Use `grid_assign_regular()` to specify that you want
# a regular sequence of size 20 across the range of cost
param_set() \%>\%
  attach_param(cost()) \%>\%
  grid_assign_regular(cost = 20) \%>\%
  grid_generate()

# Use `grid_assign_random()` to specify that you want
# 30 random uniform values across the range of cost
set.seed(123)

param_set() \%>\%
  attach_param(cost()) \%>\%
  grid_assign_random(cost = 30) \%>\%
  grid_generate()

# When there are conflicting parameters, use `id`
# and specify the name as `<name>..<id>`
param_set() \%>\%
  attach_param(cost(), id = "1") \%>\%
  attach_param(cost(), id = "2") \%>\%
  grid_assign_regular(cost..1 = 5, cost..2 = 10) \%>\%
  grid_generate()

}
\seealso{
Other parameter set functions: \code{\link{attach_param}},
  \code{\link{grid_generate}}, \code{\link{gridify}},
  \code{\link{param_set}}
}
\concept{parameter set functions}
