% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_set.R
\name{grid_generate}
\alias{grid_generate}
\title{Generate a compact parameter grid from a \code{param_set}}
\usage{
grid_generate(x, default = "regular", n = 10)
}
\arguments{
\item{x}{A \code{param_set} or \code{grid_set} to generate a \code{tbl_grid} from.}

\item{default}{A single character. Either \code{"regular"} or \code{"random"}. This is
the default type of grid to generate for parameters who's type has not
been set with \code{\link[=grid_assign_random]{grid_assign_random()}} or \code{\link[=grid_assign_regular]{grid_assign_regular()}}.}

\item{n}{A single integer. The value that goes along with \code{default}. If
\code{"regular"}, this is the number of \code{levels} for each parameter. If
\code{"random"}, this is the \code{size} for each parameter.}
}
\value{
A \code{tbl_grid} with columns for \code{name}, \code{id} and \code{values}. The \code{values} column
is a compact list column that contains the value for each hyperparameter.
}
\description{
\code{grid_generate()} constructs a \code{tbl_grid} from a \code{param_set}. This is a
compact representation of a parameter grid that can be used in other tuning
functions.
}
\details{
Remember that before a grid can be generated, none of the parameters can
contain \code{unknown()} values. These are generally set explicitly by
specifying the \code{range} when creating quantitative parameters, or by calling
\code{\link[=finalize]{finalize()}} with the training data.
}
\examples{

# Default grid, 1 param
param_set() \%>\%
  attach_param(cost()) \%>\%
  grid_generate()

# 100 random draws for the custom numeric parameter
set.seed(123)

param_set() \%>\%
  attach_param(cost()) \%>\%
  attach_range_param("custom_param", 1, 5) \%>\%
  grid_assign_random(custom_param = 100) \%>\%
  grid_generate()

# mtry() requires training data to finalize the range
mtry()

param_set() \%>\%
  attach_param(mtry()) \%>\%
  finalize(x = iris) \%>\%
  grid_generate()

}
\seealso{
Other parameter set functions: \code{\link{attach_param}},
  \code{\link{grid_assign_regular}}, \code{\link{gridify}},
  \code{\link{param_set}}
}
\concept{parameter set functions}
