% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{grid_hybrid}
\alias{grid_hybrid}
\alias{grid_hybrid.default}
\alias{grid_hybrid.parameters}
\alias{grid_hybrid.list}
\alias{grid_hybrid.param}
\title{Hybrid space-filling/regular parameter grids}
\usage{
grid_hybrid(
  x,
  ...,
  parameters = NULL,
  size = 10,
  levels = 20,
  original = TRUE,
  type = "any"
)

\method{grid_hybrid}{default}(
  x,
  ...,
  parameters = NULL,
  size = 10,
  levels = 20,
  original = TRUE,
  type = "any"
)

\method{grid_hybrid}{parameters}(
  x,
  ...,
  parameters = NULL,
  size = 10,
  levels = 20,
  original = TRUE,
  type = "any"
)

\method{grid_hybrid}{list}(
  x,
  ...,
  parameters = NULL,
  size = 10,
  levels = 20,
  original = TRUE,
  type = "any"
)

\method{grid_hybrid}{param}(
  x,
  ...,
  parameters = NULL,
  size = 10,
  levels = 20,
  original = TRUE,
  type = "any"
)
}
\arguments{
\item{x}{A \code{param} object, list, or \code{parameters}.}

\item{...}{One or more \code{param} objects (such as \code{\link[=mtry]{mtry()}} or
\code{\link[=penalty]{penalty()}}). None of the objects can have \code{unknown()} values in
the parameter ranges or values.}

\item{parameters}{A character string that matches the \emph{names} of the
parameter(s) that are used to make the regular portion of the grid. If no
value is given, a space-filling design with \code{size} candidates is created. If
all parameters are selected, a regular grid with \code{levels^p} candidates is
created where \code{p} is the length of \code{parameters}.}

\item{size}{An integer for the maximum size of the space-filling portion
of the design.}

\item{levels}{The number of values for \emph{each} regular grid parameter.}

\item{original}{A logical: should the parameters be in the original units or
in the transformed space (if any)?}

\item{type}{A character string with possible values: \code{"any"},
\code{"audze_eglais"}, \code{"max_min_l1"}, \code{"max_min_l2"}, \code{"uniform"},
\code{"max_entropy"}, or \code{"latin_hypercube"}. A value of \code{"any"} will choose the
first design available (in the order listed above, excluding
\code{"latin_hypercube"}). For a single-point design, a random grid is created.}
}
\description{
A grid can be created that is efficient for some parameter(s) but
fine-grained for others. This can be especially helpful when a model has
a "submodel" parameter that can be tuned with virtually no cost.
}
\details{
This function first creates a space-filling design for the parameters that
do not match \code{parameters} (with \code{size} total candidates). Then a regular grid
is created with \code{levels^p} candidates where \code{p} is the length of
\code{parameters}. These two grids are crossed to produce the end result.
}
\examples{
if (rlang::is_installed("ggplot2")) {

  library(dplyr)
  library(ggplot2)

  # Most boosting methods can make many predictions across a number of trees
  # from a single model fit object. Those are nearly free (computationally) so
  # we would do many tree values for each of the other parameters.

  # To illustrate, we'll only show four tree values:

  boost_example <-
    parameters(list(trees = trees(), learn_rate = learn_rate(), min_n = min_n()))

  boost_example |>
    grid_hybrid(parameters = "trees", size = 20, levels = 4) |>
    ggplot(aes(learn_rate, min_n)) +
    geom_point() +
    facet_wrap(~ trees, labeller = "label_both") +
    scale_x_log10()

  # In other cases, we have 1+ parameters with very few values. We can make a
  # small regular grid over these and a much larger space-filling design

  nnet_example <-
    parameters(
      list(
        learn_rate = learn_rate(),
        dropout = dropout(),
        # Only a few values:
        activation = activation(c("relu", "tanh", "elu")),
        schedule = rate_schedule(c("none", "cyclic", "step")) # note different name
      )
    )

  nnet_example |>
    grid_hybrid(parameters = c("activation", "schedule"), size = 20, levels = 3) |>
    ggplot(aes(learn_rate, dropout)) +
    geom_point() +
    facet_grid(activation ~ schedule, labeller = "label_both") +
    scale_x_log10()
}

}
