% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_set.R
\name{gridify}
\alias{gridify}
\title{Expand a \code{tbl_grid} into the full parameter grid}
\usage{
gridify(x)
}
\arguments{
\item{x}{A \code{tbl_grid}.}
}
\value{
A \code{tibble} with as many columns as there were rows in the \code{tbl_grid}
(one per parameter). There are as many rows as the cartesian product of
all of the parameters.
}
\description{
Most of the time, a \code{tbl_grid} suffices as the object that can be passed
off to a tuning function, but sometimes it is helpful to generate the full
grid explicitly. \code{gridify()} does exactly this.
}
\details{
When there are identical \code{name} columns in the \code{tbl_grid}, the \code{id} is used
in the column name of the expanded grid using the convention \code{<name>..<id>}
to tell them apart.
}
\examples{

# Defaults to 10 values per parameter
# 10 x 10 = 100 rows
param_set() \%>\%
  attach_param(cost()) \%>\%
  attach_param(penalty()) \%>\%
  grid_generate() \%>\%
  gridify()

# If an ID is set, and there are no conflicts,
# it is not used in the naming of the columns
param_set() \%>\%
  attach_param(cost(), id = "1") \%>\%
  grid_generate() \%>\%
  gridify()

# But when there are conflicts, the ID
# is used to tell them apart
param_set() \%>\%
  attach_param(cost(), id = "1") \%>\%
  attach_param(cost(), id = "2") \%>\%
  grid_generate() \%>\%
  gridify()

}
\seealso{
Other parameter set functions: \code{\link{attach_param}},
  \code{\link{grid_assign_regular}},
  \code{\link{grid_generate}}, \code{\link{param_set}}
}
\concept{parameter set functions}
