% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_initial_umap.R
\docType{data}
\name{initial_umap}
\alias{initial_umap}
\alias{values_initial_umap}
\title{Initialization method for UMAP}
\format{
An object of class \code{character} of length 8.
}
\usage{
initial_umap(values = values_initial_umap)

values_initial_umap
}
\arguments{
\item{values}{A character string of possible values. See \code{values_initial_umap}
in examples below.}
}
\description{
This parameter is the type of initialization for the UMAP coordinates. Can be
one of \code{"spectral"}, \code{"normlaplacian"}, \code{"random"}, \code{"lvrandom"},
\code{"laplacian"}, \code{"pca"}, \code{"spca"}, or \code{"agspectral"}. See \code{uwot::umap()} for
more details.
}
\details{
This parameter is used in \code{embed::step_umap()}.
}
\examples{
values_initial_umap
initial_umap()
}
\keyword{datasets}
