% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_set.R
\name{is_param_set}
\alias{is_param_set}
\alias{is_grid_set}
\title{Test if an object is a \code{param_set} or \code{grid_set}}
\usage{
is_param_set(x)

is_grid_set(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Test if an object is a \code{param_set} or \code{grid_set}
}
\examples{

is_param_set(param_set())

# `param_set`s become `grid_set`s after a `grid_assign_*()` function
# has been called.
param_set() \%>\%
  attach_param(cost()) \%>\%
  grid_assign_regular(cost = 5) \%>\%
  is_grid_set()

}
