% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_regular_grid.R
\name{is_regular_grid}
\alias{is_regular_grid}
\title{Determine if a grid is regular}
\usage{
is_regular_grid(grid)
}
\arguments{
\item{grid}{A tibble of grid points.}
}
\value{
A single logical
}
\description{
This function contains a set of rules that were derived from simulated data
in order to determine if a grid is regular.
}
\examples{
set_1 <- parameters(penalty(), weight_func(), learn_rate())
set_2 <- parameters(neighbors(), min_n())

is_regular_grid(grid_regular(set_1, levels = c(6, 2, 1)))
is_regular_grid(grid_regular(set_2, levels = c(10, 6)))

set.seed(13411)
is_regular_grid(nr_grid_1 <- grid_max_entropy(set_1, size = 20))
is_regular_grid(nr_grid_2 <- grid_max_entropy(set_2, size = 10))
}
\keyword{internal}
