% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_engine_mclust.R
\name{circular}
\alias{circular}
\alias{zero_covariance}
\alias{shared_orientation}
\alias{shared_shape}
\alias{shared_size}
\title{Parameters for possible engine parameters for mclust}
\usage{
circular(values = c(TRUE, FALSE))

zero_covariance(values = c(TRUE, FALSE))

shared_orientation(values = c(TRUE, FALSE))

shared_shape(values = c(TRUE, FALSE))

shared_size(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{For \code{circular()}, \code{zero_covariance()},
\code{shared_orientation()}, \code{shared_shape()}, and \code{shared_size()} either \code{TRUE} or \code{FALSE}.}
}
\description{
These parameters are used for fitting Gaussian mixture models that use the
mclust engine. They correspond to tuning parameters that would be specified using
\code{set_engine("mclust", ...)}.
}
\details{
To use these, check \code{?tidyclust::gm_clust} to see how they are used.
}
\examples{
circular()
zero_covariance()
shared_orientation()
shared_shape()
shared_size()
}
