% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.model_spec}
\alias{merge.model_spec}
\alias{merge.model_spec.tbl_grid}
\alias{merge.recipe}
\alias{merge.recipe.tbl_grid}
\alias{merge.formula}
\alias{merge.formula.tbl_grid}
\alias{merge.tbl_grid}
\alias{merge.tbl_grid.recipe}
\alias{merge.tbl_grid.model_spec}
\alias{merge.tbl_grid.formula}
\title{Merge parameter grid values into a parsnip object}
\usage{
\method{merge}{model_spec}(x, y, ...)

\method{merge.model_spec}{tbl_grid}(x, y, ...)

\method{merge}{recipe}(x, y, ...)

\method{merge.recipe}{tbl_grid}(x, y, ...)

\method{merge}{formula}(x, y, ...)

\method{merge.formula}{tbl_grid}(x, y, ...)

\method{merge}{tbl_grid}(x, y, ...)

\method{merge.tbl_grid}{recipe}(x, y, ...)

\method{merge.tbl_grid}{model_spec}(x, y, ...)

\method{merge.tbl_grid}{formula}(x, y, ...)
}
\arguments{
\item{x, y}{A combination of one \pkg{parsnip} model object (that has class
\code{"model_spec"}) and one parameter grid resulting from \code{\link[=grid_regular]{grid_regular()}} or
\code{\link[=grid_random]{grid_random()}}. As long as this combination is present, the order that
they are provided in does not matter.}

\item{...}{Not currently used.}
}
\value{
A list containing updated model objects.
}
\description{
\pkg{parsnip} contains model objects that have consistent names with
\pkg{dials}. \code{merge()} can be used to easily update any of the main parameters
in a \pkg{parsnip} model.

For more information and examples, see the article,
\href{https://tidymodels.github.io/dials/articles/articles/Dials_and_Parsnip.html}{Using dials with parsnip}.
}
