% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.recipe}
\alias{merge.recipe}
\alias{merge.model_spec}
\title{Merge parameter grid values into objects}
\usage{
\method{merge}{recipe}(x, y, ...)

\method{merge}{model_spec}(x, y, ...)
}
\arguments{
\item{x}{A recipe or model specification object.}

\item{y}{A data frame or a parameter grid resulting from one of the
\code{grid_*} functions. The column names should correspond to the parameter
names (or their annotations) in the object.}

\item{...}{Not used but required for S3 completeness.}
}
\value{
A tibble with a column \code{x} that has as many rows as were in \code{y}.
}
\description{
\code{merge()} can be used to easily update any of the arguments in a
\pkg{parsnip} model or recipe.
}
