% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_min_dist.R
\name{min_dist}
\alias{min_dist}
\title{Parameter for the effective minimum distance between embedded points}
\usage{
min_dist(range = c(-4, 0), trans = transform_log10())
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Used in \code{embed::step_umap()}.
}
\examples{
min_dist()
}
