% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_min_points.R
\name{min_points}
\alias{min_points}
\title{Minimum number of nearby points to be considered a core-point}
\usage{
min_points(range = c(3L, 50L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Used in \code{tidyclust::db_clust} model.
}
\examples{
min_points()
}
