% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_min_support.R
\name{min_support}
\alias{min_support}
\title{Parameter to control the minimum support value for frequent itemsets}
\usage{
min_support(range = c(0.1, 0.5), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Used in all \code{tidyclust::freq_itemsets()} models.
}
\examples{
min_support()
}
