% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_momentum.R
\name{momentum}
\alias{momentum}
\title{Gradient descent momentum parameter}
\usage{
momentum(range = c(0, 1), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
A useful parameter for neural network models using gradient descent
}
\examples{
momentum()
}
