% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_num_attn_blocks.R
\name{num_attn_blocks}
\alias{num_attn_blocks}
\title{Number of attention blocks}
\usage{
num_attn_blocks(range = c(1L, 6L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
The number of sequential attention blocks in the model architecture.
}
\details{
Used as a tuning parameter for \code{tabular_auto_int()} in the \code{tdl} package
when fit with the \code{brulee} engine. Each attention block consists of a
multi-head attention layer followed by feed-forward layers.
}
\examples{
num_attn_blocks()
num_attn_blocks(range = c(1L, 4L))

}
