% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_num_attn_feat.R
\name{num_attn_feat}
\alias{num_attn_feat}
\title{Number of attention features}
\usage{
num_attn_feat(range = c(8L, 64L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
The dimensionality of the feature space used in the attention mechanism
of tabular models.
}
\details{
Used as a tuning parameter for \code{tabular_auto_int()} in the \code{tdl} package
when fit with the \code{brulee} engine.
}
\examples{
num_attn_feat()
num_attn_feat(range = c(8L, 32L))

}
