% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_num_attn_heads.R
\name{num_attn_heads}
\alias{num_attn_heads}
\title{Number of attention heads}
\usage{
num_attn_heads(range = c(1L, 8L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
The number of parallel attention mechanisms (heads) in the multi-head
attention layer.
}
\details{
Used as a tuning parameter for \code{tabular_auto_int()} in the \pkg{tabular} package
when fit with the \code{brulee} engine. Multiple attention heads allow the model
to attend to different aspects of the input features simultaneously.
}
\examples{
num_attn_heads()
num_attn_heads(range = c(1L, 4L))

}
