% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_num_breaks.R
\name{num_breaks}
\alias{num_breaks}
\title{Number of cut-points for binning}
\usage{
num_breaks(range = c(2L, 10L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
This parameter controls how many bins are used when discretizing predictors.
}
\examples{
num_breaks()
}
