% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_ordinal_link.R
\name{ordinal_link}
\alias{ordinal_link}
\alias{values_ordinal_link}
\alias{odds_link}
\alias{values_odds_link}
\title{Ordinal Regression Link Functions (character)}
\usage{
ordinal_link(values = values_ordinal_link)

values_ordinal_link

odds_link(values = values_odds_link)

values_odds_link
}
\arguments{
\item{values}{For \verb{*_link()}, a character string of possible values. See
\verb{values_*_link} in examples below.}
}
\description{
The ordinal and odds link functions of an ordinal regression model.
}
\details{
These parameters are used by ordinal regression models specified by
\code{parsnip::ordinal_reg()}, for example \code{parsnip::set_engine('polr')}. The
nomenclature is taken from Wurm et al (2021), who characterize the pair of
functions as a composite link function. Note that different engines support
different subsets of link functions.
}
\examples{
values_ordinal_link
ordinal_link()
values_odds_link
odds_link()
}
\references{
Wurm, Michael J., Rathouz, Paul J., & Hanlon, Bret M. (2021).
Regularized Ordinal Regression and the ordinalNet R Package. \emph{Journal of
Statistical Software}, 99(6), 1–42. 10.18637/jss.v099.i06
}
