% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_set.R
\name{param_set}
\alias{param_set}
\title{Create a new \code{param_set}}
\usage{
param_set()
}
\description{
A \code{param_set} is the container that stores multiple param objects. These are
used to generate hyperparameter grids and other more complex
parameter structures.
}
\examples{

# An empty param_set
param_set()

# A param set with 1 param, cost()
param_set() \%>\%
  attach_param(cost())

}
\seealso{
Other parameter set functions: \code{\link{attach_param}},
  \code{\link{grid_assign_regular}},
  \code{\link{grid_generate}}, \code{\link{gridify}}
}
\concept{parameter set functions}
