% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{mtry}
\alias{mtry}
\alias{trees}
\alias{min_n}
\alias{sample_size}
\alias{learn_rate}
\alias{loss_reduction}
\alias{tree_depth}
\alias{weight_decay}
\alias{dropout}
\alias{epochs}
\alias{activation}
\alias{mixture}
\alias{regularization}
\alias{weight_func}
\alias{surv_dist}
\alias{prune}
\title{Parameter Objects}
\format{An object of class \code{quant_param} (inherits from \code{param}) of length 6.}
\usage{
mtry

trees

min_n

sample_size

learn_rate

loss_reduction

tree_depth

weight_decay

dropout

epochs

activation

mixture

regularization

weight_func

surv_dist

prune
}
\value{
Each object is generated by either \code{new_quant_param} or
\code{new_qual_param}.
}
\description{
These are objects that can be used for modeling, especially in conjunction
with the \pkg{parsnip} package.
}
\keyword{datasets}
