% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_penalty_type.R
\name{penalty_type}
\alias{penalty_type}
\alias{values_penalty_type}
\title{Regularization norm type}
\usage{
penalty_type(values = values_penalty_type)

values_penalty_type
}
\arguments{
\item{values}{A character vector of possible values.}
}
\description{
The type of regularization norm applied to per-weight coefficients in
Regularization Learning Networks.
}
\details{
Used as a tuning parameter for \code{tabular_rln()} in the \code{tdl} package when
fit with the \code{brulee} engine. L1 is recommended by the original paper.
}
\examples{
values_penalty_type
penalty_type()

}
