% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_predictor_prop.R
\name{predictor_prop}
\alias{predictor_prop}
\title{Proportion of predictors}
\usage{
predictor_prop(range = c(0, 1), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
The parameter is used in models where a parameter is the proportion of
predictor variables.
}
\details{
\code{predictor_prop()} is used in \code{step_pls()}.
}
\examples{
predictor_prop()
}
