% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_prop_terms.R
\name{prop_terms}
\alias{prop_terms}
\title{Proportion of top predictors}
\usage{
prop_terms(range = c(0.05, 1), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
The parameter is used in models where a parameter is the proportion of
predictor variables.
}
\details{
\code{prop_terms()} is used in recipe stes in the \pkg{important} package.
}
\examples{
prop_terms()
}
