% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_dials_object.R
\name{pull_dials_object}
\alias{pull_dials_object}
\alias{pull_dials_object.parameters}
\alias{pull_dials_object.recipe}
\alias{pull_dials_object.model_spec}
\alias{pull_dials_object.workflow}
\title{Return a dials parameter object associated with parameters}
\usage{
pull_dials_object(x, id, ...)

\method{pull_dials_object}{parameters}(x, id, ...)

\method{pull_dials_object}{recipe}(x, id, ...)

\method{pull_dials_object}{model_spec}(x, id, ...)

\method{pull_dials_object}{workflow}(x, id, ...)
}
\arguments{
\item{x}{The results of a call to \code{\link[=parameters]{parameters()}}, a recipe, model
specification, or workflow.}

\item{id}{A single string for the \code{id} of the parameter.}

\item{...}{Not currently used.}
}
\value{
A \code{dials} parameter object.
}
\description{
\code{pull_dials_object()} can extract a single \code{dials} parameter object from
different types of objects (e.g. parameter sets, recipes, etc.).
}
\examples{
glmn_param <- parameters(lambda = penalty(), mixture())
pull_dials_object(glmn_param, "lambda")

# fails:
# pull_dials_object(glmn_param, "penalty")
}
