% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_radius.R
\name{radius}
\alias{radius}
\title{Radius used to determine core-points and cluster assignments}
\usage{
radius(range = c(-2, 2), trans = transform_log10())
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Used in \code{\link[tidyclust:db_clust]{tidyclust::db_clust()}}.
}
\examples{
radius()
}
