% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-scales.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{log10_trans}
\alias{log2_trans}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{scales}{\code{\link[scales]{log10_trans}}, \code{\link[scales]{log2_trans}}}
}}

