% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_regularization_method.R
\name{regularization_method}
\alias{regularization_method}
\alias{values_regularization_method}
\title{Estimation methods for regularized models}
\usage{
regularization_method(values = values_regularization_method)

values_regularization_method
}
\arguments{
\item{values}{A character string of possible values. See \code{values_regularization_method}
in examples below.}
}
\description{
Estimation methods for regularized models
}
\details{
This parameter is used in \code{parsnip::discrim_linear()}.
}
\examples{
values_regularization_method
regularization_method()
}
