% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_resid_at.R
\name{resid_at}
\alias{resid_at}
\title{Residual connection locations}
\usage{
resid_at(range = c(2L, unknown()), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
The layer indices at which residual connections are added in a tabular
residual network.
}
\details{
Used as a tuning parameter for \code{tabular_resnet()} in the \pkg{tabular} package
when fit with the \code{brulee} engine. The upper bound depends on the number
of hidden layers in the network and so is left as \code{unknown()} by default.
}
\examples{
resid_at()
resid_at(range = c(2L, 6L))

}
