% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_resnet.R
\name{resnet-param}
\alias{resnet-param}
\alias{bottleneck_units}
\alias{resid_at}
\title{Parameters for residual networks}
\usage{
bottleneck_units(range = c(2L, 25L), trans = NULL)

resid_at(range = c(2L, unknown()), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
These parameters are used with tabular residual networks, such as
\code{tabular_resnet()} when fit with the \code{brulee} engine.
}
\details{
\itemize{
\item \code{bottleneck_units()}: The number of units used in ResNet layers within a
residual block.
\item \code{resid_at()}: The layer indices at which residual connections are added.
The upper bound depends on the number of hidden layers in the network and
so is left as \code{unknown()} by default.
}
}
\examples{
bottleneck_units()
resid_at()
}
