% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_rln.R
\name{rln-param}
\alias{rln-param}
\alias{penalty_average}
\alias{step_rate}
\alias{penalty_type}
\alias{values_penalty_type}
\title{Parameters for regularization learning networks}
\usage{
penalty_average(range = c(-15, -5), trans = scales::transform_log10())

step_rate(range = c(0, 8), trans = scales::transform_log10())

penalty_type(values = values_penalty_type)

values_penalty_type
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}

\item{values}{A character vector of possible values.}
}
\description{
These parameters are used with regularization learning networks (RLN), such
as \code{tabular_rln()} when fit with the \code{brulee} engine.
}
\details{
\itemize{
\item \code{penalty_average()}: The target geometric mean of the per-weight
regularization coefficients (Theta in Shavitt and Segal (2018)). Best tuned
on the log10 scale.
\item \code{step_rate()}: The step size used to update the per-weight regularization
coefficients (nu in Shavitt and Segal (2018)). Best tuned on the log10
scale.
\item \code{penalty_type()}: The type of regularization norm applied to per-weight
coefficients. L1 is recommended by the original paper.
}

For \code{penalty_average()} and \code{step_rate()}, the value is passed to
\code{brulee::brulee_rln()} on the natural scale.
}
\examples{
penalty_average()
step_rate()
values_penalty_type
penalty_type()
}
\references{
Shavitt, I., & Segal, E. (2018). Regularization learning networks: Deep
learning for tabular datasets. \emph{Advances in Neural Information Processing
Systems}, 31, 1379-1389.
}
