% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_select_features.R
\name{select_features}
\alias{select_features}
\title{Parameter to enable feature selection}
\usage{
select_features(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{A vector of possible values (TRUE or FALSE).}
}
\description{
Used in \code{parsnip::gen_additive_mod()}.
}
\examples{
select_features()
}
