% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_shared_size.R
\name{shared_size}
\alias{shared_size}
\title{Parameter to control whether fitted Gaussian distributions for have clusters share the same size in the predictor space}
\usage{
shared_size(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{A vector of possible values (\code{TRUE} or \code{FALSE}).}
}
\description{
Used in \code{\link[tidyclust:gm_clust]{tidyclust::gm_clust()}}.
}
\examples{
shared_size()
}
