% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_step_rate.R
\name{step_rate}
\alias{step_rate}
\title{Step size for per-weight regularization updates}
\usage{
step_rate(range = c(0, 8), trans = scales::transform_log10())
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
The step size used to update the per-weight regularization coefficients
(nu in Shavitt and Segal (2018)) for Regularization Learning Networks.
Best tuned on the log10 scale.
}
\details{
Used as a tuning parameter for \code{tabular_rln()} in the \pkg{tabular} package when
fit with the \code{brulee} engine. The value is passed to \code{brulee::brulee_rln()}
on the natural scale.
}
\examples{
step_rate()
step_rate(range = c(2, 6))

}
\references{
Shavitt, I., & Segal, E. (2018). Regularization learning networks: Deep
learning for tabular datasets. \emph{Advances in Neural Information Processing
Systems}, 31, 1379-1389.
}
