% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_surv_dist.R
\name{surv_dist}
\alias{surv_dist}
\alias{values_surv_dist}
\title{Parametric distributions for censored data}
\usage{
surv_dist(values = values_surv_dist)

values_surv_dist
}
\arguments{
\item{values}{A character string of possible values. See \code{values_surv_dist}
in examples below.}
}
\description{
Parametric distributions for censored data
}
\details{
This parameter is used in \code{parsnip::survival_reg()}.
}
\examples{
values_surv_dist
surv_dist()
}
