% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_surv_scale.R
\docType{data}
\name{surv_link}
\alias{surv_link}
\alias{values_surv_link}
\title{Survival Model Link Function}
\format{
An object of class \code{character} of length 3.
}
\usage{
surv_link(values = values_surv_link)

values_surv_link
}
\arguments{
\item{values}{A character string of possible values. See \code{values_surv_link}
in examples below.}
}
\description{
Survival Model Link Function
}
\details{
This parameter is used in \code{parsnip::set_engine('flexsurvspline')}.
}
\examples{
values_surv_link
surv_link()
}
\keyword{datasets}
