% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_surv_scale.R
\docType{data}
\name{surv_scale}
\alias{surv_scale}
\alias{values_surv_scale}
\title{Survival Model Scale}
\format{
An object of class \code{character} of length 3.
}
\usage{
surv_scale(values = values_surv_scale)

values_surv_scale
}
\arguments{
\item{values}{A character string of possible values. See \code{values_surv_scale}
in examples below.}
}
\description{
Survival Model Scale
}
\details{
This parameter is used in \code{parsnip::set_engine('flexsurvspline')}.
}
\examples{
values_surv_scale
surv_scale()
}
\keyword{datasets}
