% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_tab_pfn.R
\name{tab-pfn-param}
\alias{tab-pfn-param}
\alias{num_estimators}
\alias{softmax_temperature}
\alias{balance_probabilities}
\alias{average_before_softmax}
\alias{training_set_limit}
\title{Parameters for TabPFN models
These parameters are used for constructing Prior data fitted network (TabPFN)
models.}
\usage{
num_estimators(range = c(1, 25), trans = NULL)

softmax_temperature(range = c(0, 10), trans = NULL)

balance_probabilities(values = c(TRUE, FALSE))

average_before_softmax(values = c(TRUE, FALSE))

training_set_limit(range = c(2L, 10000L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}

\item{values}{A vector of possible values (TRUE or FALSE).}
}
\description{
Parameters for TabPFN models
These parameters are used for constructing Prior data fitted network (TabPFN)
models.
}
\details{
These parameters are often used with TabPFN models via \code{parsnip::tab_pfn()}.
}
