% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_target_weight.R
\name{target_weight}
\alias{target_weight}
\title{Amount of supervision parameter}
\usage{
target_weight(range = c(0, 1), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
For \code{uwot::umap()} and \code{embed::step_umap()}, this is a weighting factor
between data topology and target topology. A value of 0.0 weights entirely
on data, a value of 1.0 weights entirely on target. The default of 0.5
balances the weighting equally between data and target.
}
\details{
This parameter is used in \code{recipes} via \code{embed::step_umap()}.
}
\examples{
target_weight()
}
