% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_token.R
\docType{data}
\name{token}
\alias{token}
\alias{values_token}
\title{Token types}
\format{
An object of class \code{character} of length 12.
}
\usage{
token(values = values_token)

values_token
}
\arguments{
\item{values}{A character string of possible values. See \code{values_token}
in examples below.}
}
\description{
Token types
}
\details{
This parameter is used in \code{textrecipes::step_tokenize()}.
}
\examples{
values_token
token()
}
\keyword{datasets}
