% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_trim_amount.R
\name{trim_amount}
\alias{trim_amount}
\title{Amount of Trimming}
\usage{
trim_amount(range = c(0, 0.5), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Used in \code{recipes::step_impute_mean()}.
}
\examples{
trim_amount()
}
