% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_sum.R
\name{type_sum.param}
\alias{type_sum.param}
\title{Succinct summary of parameter objects}
\usage{
\method{type_sum}{param}(x)
}
\arguments{
\item{x}{A \code{param} object to summarise.}
}
\value{
A character value.
}
\description{
\code{type_sum} controls how objects are shown when inside tibble
columns.
}
\details{
For \code{param} objects, the summary prefix is either
"\code{dparam}" (if a qualitative parameter) or "\code{nparam}" (if
quantitative). Additionally, brackets are used to indicate if
there are unknown values. For example, "\code{nparam[?]}" would
indicate that part of the numeric range is has not been
finalized and "\code{nparam[+]}" indicates a parameter that is
complete.
}
\keyword{internal}
