% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_unknown.R
\name{unknown}
\alias{unknown}
\alias{is_unknown}
\alias{has_unknowns}
\title{Placeholder for Unknown Parameter Values}
\usage{
unknown()

is_unknown(x)

has_unknowns(object)
}
\arguments{
\item{x}{An object or vector or objects.}

\item{object}{An object of class \code{param}}
}
\value{
\code{unknown} returns expression value for \code{unknown()} and logicals for
\code{is_unknown()} and \code{has_unknowns()}.
}
\description{
This creates a simple expression used to signify that the value will be
specified at a later time.
}
