% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_set.R
\name{update.param_set}
\alias{update.param_set}
\title{Update a single parameter in a parameter set}
\usage{
\method{update}{param_set}(object, id, value, ...)
}
\arguments{
\item{object}{A parameter set.}

\item{id}{A single character value that will be used in a regular expression
to select a single parameter for updating.}

\item{value}{A \code{param} object or \code{NA}}

\item{...}{Options to pass to \code{grep()}}
}
\value{
The modified parameter set.
}
\description{
Update a single parameter in a parameter set
}
\examples{
params <- list(lambda = penalty(), alpha = mixture(), `rand forest` = mtry())
pset <- param_set(params)
pset

update(pset, "forest", finalize(mtry(), iris))
}
