% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_weight_func.R
\docType{data}
\name{weight_func}
\alias{weight_func}
\alias{values_weight_func}
\title{Kernel functions for distance weighting}
\format{
An object of class \code{character} of length 9.
}
\usage{
weight_func(values = values_weight_func)

values_weight_func
}
\arguments{
\item{values}{A character string of possible values. See \code{values_weight_func}
in examples below.}
}
\description{
Kernel functions for distance weighting
}
\details{
This parameter is used in \code{parsnip:::nearest_neighbors()}.
}
\examples{
values_weight_func
weight_func()
}
\keyword{datasets}
