% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_weight_scheme.R
\name{weight_scheme}
\alias{weight_scheme}
\alias{values_weight_scheme}
\title{Term frequency weighting methods}
\usage{
weight_scheme(values = values_weight_scheme)

values_weight_scheme
}
\arguments{
\item{values}{A character string of possible values. See \code{values_weight_scheme}
in examples below.}
}
\description{
Term frequency weighting methods
}
\details{
This parameter is used in \code{textrecipes::step_tf()}.
}
\examples{
values_weight_scheme
weight_scheme()
}
