% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_zero_covariance.R
\name{zero_covariance}
\alias{zero_covariance}
\title{Parameter to control whether clusters have zero covariance between predictors}
\usage{
zero_covariance(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{A vector of possible values (TRUE or FALSE).}
}
\description{
Used in \code{\link[tidyclust:gm_clust]{tidyclust::gm_clust()}}.
}
\examples{
zero_covariance()
}
