% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_activation.R
\docType{data}
\name{activation}
\alias{activation}
\alias{activation_2}
\alias{values_activation}
\title{Activation functions between network layers}
\format{
An object of class \code{character} of length 23.
}
\usage{
activation(values = values_activation)

activation_2(values = values_activation)

values_activation
}
\arguments{
\item{values}{A character string of possible values. See \code{values_activation}
in examples below.}
}
\description{
Activation functions between network layers
}
\details{
This parameter is used in \code{parsnip} models for neural networks such as
\code{parsnip:::mlp()}.
}
\examples{
values_activation
activation()
}
\keyword{datasets}
