% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_circular.R
\name{circular}
\alias{circular}
\title{Parameter to control circular or ellipsoidal cluster shapes}
\usage{
circular(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{A vector of possible values (TRUE or FALSE).}
}
\description{
Used in \code{\link[tidyclust:gm_clust]{tidyclust::gm_clust()}}.
}
\examples{
circular()
}
