% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge.model_spec}
\alias{merge.model_spec}
\title{Merge parameter grid values into a parsnip object}
\usage{
\method{merge}{model_spec}(x, y, ...)
}
\arguments{
\item{x}{A \pkg{parsnip} model object (that has class \code{model_spec}).}

\item{y}{A parameter grid resulting from \code{regular_grid} or \code{random_grid}.}

\item{...}{Not currently used.}
}
\value{
An updated model object.
}
\description{
\pkg{parsnip} contains model objects that have consistent names with
\pkg{dials}. \code{merge} can be used to easily update any of the main parameters
in a \pkg{parsnip} model.
}
