% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_min_unique.R
\name{min_unique}
\alias{min_unique}
\title{Number of unique values for pre-processing}
\usage{
min_unique(range = c(5L, 15L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Some pre-processing parameters require a minimum number of unique data points
to proceed.
}
\examples{
min_unique()
}
