% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{weight_func}
\alias{weight_func}
\alias{misc_parameters}
\alias{surv_dist}
\alias{Laplace}
\alias{neighbors}
\title{Parameter objects related to miscellaneous models.}
\format{An object of class \code{qual_param} (inherits from \code{param}) of length 4.}
\usage{
weight_func

surv_dist

Laplace

neighbors
}
\value{
Each object is generated by either \code{new_quant_param} or
\code{new_qual_param}.
}
\description{
These are objects that can be used for modeling, especially in conjunction
with the \pkg{parsnip} package.
}
\details{
These objects are pre-made parameter sets that are useful in a variety of
models.
\itemize{
\item \code{weight_func}: The type of kernel function that weights the distances
between samples (e.g. in a K-near neighbors model).
\item \code{surv_dist}: the statistical distribution of the data in a survival
analysis model (e.g. \code{parsnip::surv_reg()}) .
\item \code{Laplace}: the Laplace correction used to smooth low-frequency counts.
\item \code{neighbors}: a parameter for the number of neighbors used in a prototype
model.
}
}
\keyword{datasets}
