% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{weight_func}
\alias{weight_func}
\alias{misc_parameters}
\alias{surv_dist}
\alias{Laplace}
\alias{neighbors}
\alias{dist_power}
\alias{threshold}
\title{Parameter functions related to miscellaneous models.}
\usage{
weight_func(values = values_weight_func)

surv_dist(values = values_surv_dist)

Laplace(range = c(0, 3), trans = NULL)

neighbors(range = c(1L, unknown()), trans = NULL)

dist_power(range = c(1, 2), trans = NULL)

threshold(range = c(0, 1), trans = NULL)
}
\arguments{
\item{values}{A vector of possible parameter values. For qualitative
parameters, character and logical defaults have been set, but can be modified
as needed.}

\item{range}{A two-element vector holding the smallest and largest possible
values, respectively. If these cannot be set when the parameter is created,
\code{unknown()} can be used instead of a particular value. If a transformation
is specified, these should be in the \emph{transformed units}. The default is
an educated guess at the range of each parameter.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\value{
A \code{param} object. The underlying functions that generate the parameters are
\code{\link[=new_quant_param]{new_quant_param()}} and \code{\link[=new_qual_param]{new_qual_param()}}.
}
\description{
These are parameter generating functions that can be used for modeling,
especially in conjunction with the \pkg{parsnip} package.
}
\details{
These functions generate parameters that are useful in a variety of
models.
\itemize{
\item \code{weight_func()}: The type of kernel function that weights the distances
between samples (e.g. in a K-nearest neighbors model).
\item \code{surv_dist()}: The statistical distribution of the data in a survival
analysis model. (See \code{parsnip::surv_reg()}) .
\item \code{Laplace()}: The Laplace correction used to smooth low-frequency counts.
\item \code{neighbors()}: A parameter for the number of neighbors used in a prototype
model.
\item \code{dist_power()}: The order parameter used in calculating a
Minkowski distance.
\item \code{threshold()}: A general thresholding parameter for values
between \code{[0, 1]}.
}
}
