% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_num_knots.R
\name{num_knots}
\alias{num_knots}
\title{Number of knots (integer)}
\usage{
num_knots(range = c(0L, 5L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
The number of knots used for spline model parameters.
}
\details{
One context in which this parameter is used is spline basis functions.
}
\examples{
num_knots()
}
