% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_ordinal_link.R
\docType{data}
\name{ordinal_link}
\alias{ordinal_link}
\alias{values_ordinal_link}
\title{Ordinal Regression Link Function (character)}
\format{
An object of class \code{character} of length 5.
}
\usage{
ordinal_link(values = values_ordinal_link)

values_ordinal_link
}
\arguments{
\item{values}{A character string of possible values. See
\code{values_ordinal_link} in examples below.}
}
\description{
The link function of an ordinal (generalized linear) regression model.
}
\details{
This parameter is used in \code{parsnip::set_engine('polr')}.
}
\examples{
values_ordinal_link
ordinal_link()
}
\keyword{datasets}
