% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_set.R
\name{param_set}
\alias{param_set}
\alias{param_set.default}
\alias{param_set.param}
\alias{param_set.list}
\title{Information on tuning parameters within an object}
\usage{
param_set(x, ...)

\method{param_set}{default}(x, ...)

\method{param_set}{param}(x, ...)

\method{param_set}{list}(x, ...)
}
\arguments{
\item{x}{An object, such as a list of \code{param} objects or an actual \code{param}
object.}

\item{...}{Only used for the \code{param} method so that multiple \code{param} objects
can be passed to the function.}
}
\description{
Information on tuning parameters within an object
}
