% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\alias{parameters.default}
\alias{parameters.param}
\alias{parameters.list}
\title{Return a parameter set from one or more dials \code{param} objects.}
\usage{
parameters(x, ...)

\method{parameters}{default}(x, ...)

\method{parameters}{param}(x, ...)

\method{parameters}{list}(x, ...)
}
\arguments{
\item{x}{An object, such as a list of \code{param} objects or an actual \code{param}
object.}

\item{...}{Only used for the \code{param} method so that multiple \code{param} objects
can be passed to the function.}
}
\value{
A tibble of class \code{parameters} where each row represents one tunable
parameter.
}
\description{
Return a parameter set from one or more dials \code{param} objects.
}
\examples{
params <- list(lambda = penalty(), alpha = mixture(), `rand forest` = mtry())
pset <- parameters(params)
pset

}
\seealso{
\code{\link[=update.parameters]{update.parameters()}}, \code{\link[=finalize]{finalize()}},
\code{\link[hardhat:hardhat-extract]{hardhat::extract_parameter_set_dials()}}
}
