% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\alias{parameters.default}
\alias{parameters.param}
\alias{parameters.list}
\title{Information on tuning parameters within an object}
\usage{
parameters(x, ...)

\method{parameters}{default}(x, ...)

\method{parameters}{param}(x, ...)

\method{parameters}{list}(x, ..., .constraint = NULL)
}
\arguments{
\item{x}{An object, such as a list of \code{param} objects or an actual \code{param}
object.}

\item{...}{Only used for the \code{param} method so that multiple \code{param} objects
can be passed to the function.}
}
\description{
Information on tuning parameters within an object
}
