% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_prune_method.R
\docType{data}
\name{prune_method}
\alias{prune_method}
\alias{values_prune_method}
\title{MARS pruning methods}
\format{
An object of class \code{character} of length 6.
}
\usage{
prune_method(values = values_prune_method)

values_prune_method
}
\arguments{
\item{values}{A character string of possible values. See \code{values_prune_method}
in examples below.}
}
\description{
MARS pruning methods
}
\details{
This parameter is used in \code{parsnip:::mars()}.
}
\examples{
values_prune_method
prune_method()
}
\keyword{datasets}
