% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_dials_object.R
\name{pull_dials_object}
\alias{pull_dials_object}
\title{Return a dials parameter object associated with parameters}
\usage{
pull_dials_object(x, id, ...)
}
\arguments{
\item{x}{The results of a call to \code{\link[=parameters]{parameters()}}, a recipe, model
specification, or workflow.}

\item{id}{A single string for the \code{id} of the parameter.}

\item{...}{Not currently used.}
}
\value{
A \code{dials} parameter object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been deprecated; please use
\code{hardhat::extract_parameter_dials()} instead.

\code{pull_dials_object()} can extract a single \code{dials} parameter object from
different types of objects (e.g. parameter sets, recipes, etc.).
}
\examples{
glmn_param <- parameters(lambda = penalty(), mixture())
# pull_dials_object(glmn_param, "lambda")
# ->
extract_parameter_dials(glmn_param, "lambda")
}
\keyword{internal}
