% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_shared_orientation.R
\name{shared_orientation}
\alias{shared_orientation}
\title{Parameter to control whether clusters share the same orientation in the predictor space
(Only applicable if zero_covariance = FALSE)}
\usage{
shared_orientation(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{A vector of possible values (TRUE or FALSE).}
}
\description{
Used in \code{tidyclust::gm_clust()}.
}
\examples{
shared_orientation()
}
