% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_shared_shape.R
\name{shared_shape}
\alias{shared_shape}
\title{Parameter to control whether clusters share the same shape in the predictor space
(Only applicable if circular = FALSE)}
\usage{
shared_shape(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{A vector of possible values (TRUE or FALSE).}
}
\description{
Used in \code{tidyclust::gm_clust()}.
}
\examples{
shared_shape()
}
