% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_summary_stat.R
\docType{data}
\name{summary_stat}
\alias{summary_stat}
\alias{values_summary_stat}
\title{Rolling summary statistic for moving windows}
\format{
An object of class \code{character} of length 8.
}
\usage{
summary_stat(values = values_summary_stat)

values_summary_stat
}
\arguments{
\item{values}{A character string of possible values. See \code{values_summary_stat}
in examples below.}
}
\description{
This parameter is used in \code{recipes::step_window()}.
}
\examples{
values_summary_stat
summary_stat()
}
\keyword{datasets}
