% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_survival_link.R
\docType{data}
\name{survival_link}
\alias{survival_link}
\alias{values_survival_link}
\title{Survival Model Link Function}
\format{
An object of class \code{character} of length 3.
}
\usage{
survival_link(values = values_survival_link)

values_survival_link
}
\arguments{
\item{values}{A character string of possible values.
See \code{values_survival_link} in examples below.}
}
\description{
Survival Model Link Function
}
\details{
This parameter is used in \code{parsnip::set_engine('flexsurvspline')}.
}
\examples{
values_survival_link
survival_link()
}
\keyword{datasets}
