% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_unknown.R
\name{unknown}
\alias{unknown}
\alias{is_unknown}
\title{Placeholder for Unknown Parameter Values}
\usage{
unknown()

is_unknown(x)
}
\arguments{
\item{x}{An object.}
}
\value{
An expression value for \code{unknown()} and a logical for \code{is_unknown()}.
}
\description{
This creates a simple expression used to signify that the value will be
specified at a later time.
}
