% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{value_validate}
\alias{value_validate}
\alias{value_seq}
\alias{value_sample}
\alias{value_transform}
\alias{value_inverse}
\title{Tools for working with parameter values}
\usage{
value_validate(object, values)

value_seq(object, n, original = TRUE)

value_sample(object, n, original = TRUE)

value_transform(object, values)

value_inverse(object, values)
}
\arguments{
\item{object}{An object with class \code{quant_param}.}

\item{values}{A numeric vector or list (including \code{Inf}). Values
\emph{cannot} include \code{unknown()}.}

\item{n}{An integer for the number of values to return.}

\item{original}{A single logical: should the range values be in the natural
units (\code{TRUE}) or in the transformed space (\code{FALSE}, if applicable).}
}
\value{
\code{value_validate()} throws an error or silently returns the values.
\code{value_transform} and \code{value_inverse} return a \emph{vector} of numeric values
while \code{value_seq} and \code{value_sample} return a vector of values consistent
with the \code{type} field of \code{object}.
}
\description{
Tools for working with parameter values
}
