% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_shared_orientation.R,
%   R/param_zero_covariance.R
\name{zero_covariance}
\alias{zero_covariance}
\title{Parameter to control whether clusters share the same orientation in the predictor space
(Only applicable if zero_covariance = FALSE)}
\usage{
zero_covariance(values = c(TRUE, FALSE))

zero_covariance(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{A vector of possible values (TRUE or FALSE).}
}
\description{
Used in \code{tidyclust::gm_clust()}.

Used in \code{tidyclust::gm_clust()}.
}
\examples{
shared_orientation()
zero_covariance()
}
